todo_project/requirements.txt
django
djangorestframework
psycopg2-binary
celery
redis
django-environ